migrator: create Eitems with files
* refactor code per rectype
