+++ cds_ils/migrator/loans/api.py
@@ -1,0 +1,130 @@
+# -*- coding: utf-8 -*-
+#
+# This file is part of Invenio.
+# Copyright (C) 2019-2020 CERN.
+#
+# CDS-ILS is free software; you can redistribute it and/or modify it
+# under the terms of the MIT License; see LICENSE file for more details.
+
+"""CDS-ILS migrator API."""
+
+import json
+import logging
+
+import click
+from elasticsearch import VERSION as ES_VERSION
+from invenio_app_ils.documents.api import Document
+from invenio_app_ils.items.api import ITEM_PID_TYPE
+from invenio_app_ils.patrons.api import SystemAgent
+from invenio_app_ils.proxies import current_app_ils
+from invenio_db import db
+
+from cds_ils.migrator.api import model_provider_by_rectype, import_record
+from cds_ils.migrator.errors import (
+    ItemMigrationError,
+    LoanMigrationError,
+)
+from cds_ils.migrator.items.api import get_item_by_barcode
+from cds_ils.migrator.patrons.api import get_user_by_legacy_id
+
+lt_es7 = ES_VERSION[0] < 7
+migrated_logger = logging.getLogger(
+                            "migrated_documents"
+                        )
+
+
+def import_loans_from_json(dump_file):
+    """Imports loan objects from JSON."""
+    dump_file = dump_file[0]
+    loans = []
+    with click.progressbar(json.load(dump_file)) as bar:
+        for record in bar:
+            click.echo('Importing loan "{0}"...'.format(record["legacy_id"]))
+            user = get_user_by_legacy_id(record["id_crcBORROWER"])
+            if not user:
+                patron_pid = SystemAgent.id
+            else:
+                patron_pid = user.pid
+            try:
+                item = get_item_by_barcode(record["item_barcode"])
+            except ItemMigrationError:
+                continue
+                # Todo uncomment when more data
+                # raise LoanMigrationError(
+                #    'no item found with the barcode {} for loan {}'.format(
+                #        record['item_barcode'], record['legacy_id']))
+
+            # additional check if the loan refers to the same document
+            # as it is already attached to the item
+            document_pid = item.get("document_pid")
+            document = Document.get_record_by_pid(document_pid)
+            if record["legacy_document_id"] is None:
+                raise LoanMigrationError(
+                    "no document id for loan {}".format(record["legacy_id"])
+                )
+            if (
+                document.get("legacy_recid", None)
+                != record["legacy_document_id"]
+            ):
+                # this might happen when record merged or migrated,
+                # the already migrated document should take precedence
+                click.secho(
+                    "inconsistent document dependencies for loan {}".format(
+                        record["legacy_id"]
+                    ),
+                    fg="blue",
+                )
+
+            # create a loan
+            (
+                default_location_pid_value,
+                _,
+            ) = current_app_ils.get_default_location_pid
+
+            if record["status"] == "on loan":
+                loan_dict = dict(
+                    patron_pid=str(patron_pid),
+                    transaction_location_pid=default_location_pid_value,
+                    transaction_user_pid=str(SystemAgent.id),
+                    document_pid=document_pid,
+                    item_pid={
+                        "type": ITEM_PID_TYPE,
+                        "value": item.pid.pid_value,
+                    },
+                    start_date=record["start_date"],
+                    end_date=record["end_date"],
+                    state="ITEM_ON_LOAN",
+                    transaction_date=record["start_date"],
+                )
+            elif record["status"] == "returned":
+                loan_dict = dict(
+                    patron_pid=str(patron_pid),
+                    transaction_location_pid=default_location_pid_value,
+                    transaction_user_pid=str(SystemAgent.id),
+                    transaction_date=record["returned_on"],
+                    document_pid=document_pid,
+                    item_pid={
+                        "type": ITEM_PID_TYPE,
+                        "value": item.pid.pid_value,
+                    },
+                    start_date=record["start_date"],
+                    end_date=record["returned_on"],
+                    state="ITEM_RETURNED",
+                )
+            else:
+                raise LoanMigrationError(
+                    "Unkown loan state for record {0}: {1}".format(
+                        record["legacy_id"], record["state"]
+                    )
+                )
+            model, provider = model_provider_by_rectype("loan")
+            try:
+                loan = import_record(loan_dict, model, provider)
+            except Exception as e:
+                raise e
+            try:
+                # without this script is very slow
+                db.session.commit()
+            except Exception:
+                db.session.rollback()
+    return loans
