+++ cds_ils/migrator/relations/api.py
@@ -1,0 +1,18 @@
+import click
+from invenio_app_ils.records_relations.api import RecordRelationsParentChild
+
+
+def create_parent_child_relation(parent, child, relation_type, volume):
+    """Create parent child relations."""
+    rr = RecordRelationsParentChild()
+    click.echo(
+        "Creating relations: {0} - {1}".format(parent["pid"], child["pid"])
+    )
+    rr.add(
+        parent=parent,
+        child=child,
+        relation_type=relation_type,
+        volume=str(volume) if volume else None,
+    )
+
+
