+++ scripts/life_data_setup
@@ -14,14 +14,14 @@ invenio fixtures pages
 invenio fixtures location
 invenio fixtures vocabularies
 invenio fixtures demo-patrons
+invenio files location --default ils /tmp/ils-files
 
 # This part requires mounted EOS volume or access to the files
-# invenio migration internal-locations /Volumes/cds/test/books/migration/records/tmp/libraries.json
-# invenio migration documents /Volumes/cds/test/books/migration/records/tmp/demo-books_0.json
-# invenio migration items /Volumes/cds/test/books/migration/records/tmp/items_0.json
 # cds-ils migration documents /eos/media/cds/test/books/migration/records/tmp/demo-books_0.json
 # cds-ils migration internal-locations /eos/media/cds/test/books/migration/records/tmp/libraries.json
 # cds-ils migration items /eos/media/cds/test/books/migration/records/tmp/items_0.json
 # cds-ils migration borrowers /eos/media/cds/test/books/migration/records/tmp/active_users.json
 
+# create eitems from document files
+# invenio migration eitems-files
