+++ ui/.env
@@ -1,5 +1,5 @@
 REACT_APP_UI_ILS_VERSION=1.0.0-alpha.13
 REACT_APP_INVENIO_VERSION=1.0.0a13
-REACT_APP_INVENIO_UI_URL=https://localhost:5000
-REACT_APP_INVENIO_REST_ENDPOINTS_BASE_URL=https://localhost:5000/api
+REACT_APP_INVENIO_UI_URL=https://127.0.0.1:5000
+REACT_APP_INVENIO_REST_ENDPOINTS_BASE_URL=https://127.0.0.1:5000/api
 REACT_APP_OVERLAY_VERSION=1.0.0a11
